Whole new/clean structure with lots of Documentation. With out Update
