+++ OOP_Assignment/src/Trident/Course.java
@@ -1,6 +1,5 @@
 package Trident;
 
-
 public class Course {
     String CID;
     int CreditHr;
