+++ OOP_Assignment/src/Trident/CourseTaken.java
@@ -1,6 +1,6 @@
 package Trident;
 
-public class CourseTaken extends Course{
+public class CourseTaken extends Course {
     String GradeLetter;
     public CourseTaken(String CID, int CreditHr, int ContactHr, int LabHr, int LectureHr, String CName, int CDID, String gL) {
         super(CID, CreditHr, ContactHr, LabHr, LectureHr, CName, CDID);
