--- OOP_Assignment/src/com/company/DeleteDepartment.java
@@ -1,6 +1,0 @@
-package com.company;
-
-public class DeleteDepartment {
-    public void delete() {
-    }
-}
