--- OOP_Assignment/src/com/company/Main.java
@@ -1,10 +1,0 @@
-package com.company;
-
-import static java.lang.System.*;
-
-public class Main{
-    public static void main(String[] args) {
-        Menu show = new Menu();
-        show.welcome();
-    }
-}
